==> 2_preprocessing.py <==
# -*- coding: utf-8 -*-
"""2.PreProcessing

Automatically generated by Colaboratory.

Original file is located at
    https://colab.research.google.com/drive/15PqB2hiAMM2TTu-1sXjpr9042FWnDW1T
"""

import pandas as pd
pd.set_option('Chained_assignment',None)

df=pd.read_csv("/content/drive/MyDrive/Statistical-Time-Series-Analysis-on-Agricultural-Commodity-Prices-master/Monthly_data_cmo_filtered.csv")
df['date']=pd.to_datetime(df["date"])
df['quarter']=df["date"].dt.quarter
df['Combination_label']=df["APMC"]+":"+df["Commodity"]
df.shape

def group(df):
    data=df[["APMC","Commodity","modal_price","Year","Combination_label"]].groupby(["APMC","Commodity","Year",
            "Combination_label"],as_index=False).count().rename(columns={"modal_price":"Count"}).reset_index(drop=True)
    return data

df.columns

data = group(df)
data = data[data['Count']==12].reset_index(drop=True)
data

filtered=data["Combination_label"].unique().tolist()
df_filtered=df[df["Combination_label"].isin(filtered)].reset_index(drop=True)

df_filtered.to_csv("seasonal_data_analysis.csv",index=False)
